% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmaqshiny_run.R
\name{mmaqshiny_run}
\alias{mmaqshiny_run}
\title{mmaqshiny: R-Shiny package to explore air-pollution mobile monitoring data}
\usage{
mmaqshiny_run()
}
\description{
The R-Shiny package we present here is for analysing, visualising and spatial-mapping of high-resolution air-pollution data collected by specific devices installed on a moving platform.
}
\examples{
\dontrun{
library(mmaqshiny)
mmaqshiny::mmaqshiny_run()
}
}
\keyword{air}
\keyword{mmaqshiny,}
\keyword{mobile}
\keyword{monitoring}
\keyword{quality,}
\keyword{shiny,}
