% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmaqshiny_run.R
\name{mmaqshiny_run}
\alias{mmaqshiny_run}
\title{mmaqshiny: Explore Air Quality Mobile Monitoring Data}
\usage{
mmaqshiny_run()
}
\description{
The package is for analysing, visualising and spatial mapping of
high-resolution air quality data collected by specific devices installed on
a moving platform.
}
\examples{
\dontrun{
library(mmaqshiny)
mmaqshiny::mmaqshiny_run()
}
}
\keyword{mmaqshiny}
